refactor: split fleet.rs God Class into module directory (#1375)
Convert fleet.rs (4028 lines, 6 responsibilities) into fleet/ module:

- fleet/resolve.rs: resolve_instance (60 lines) + expand_tilde_path
  utility (eliminates 3x12-line duplication)
- fleet/merge.rs: merge_instance_into_existing (39 lines) + FieldClass
  enum + merge_string_field/merge_typed_field helpers
- fleet/persist.rs: all YAML mutation functions (add/remove/update
  instance/team, migrate_teams_json_to_yaml, mutate_fleet_yaml)
- fleet/mod.rs: FleetConfig (load/parse/cache/normalize/backfill) + types

Production code in mod.rs: 588 lines. Total mod.rs: 2093 lines (tests).
resolve_instance: 60 lines (was 150). merge_instance_into_existing: 39
lines (was 143). Pure extract-method refactor — zero behavior change.


Agend-Agent: fixup-dev
Agend-Branch: refactor/1375-split-fleet
Agend-Issued-At: 2026-05-28T09:20:56.392607+00:00
